add simple error handling for brands

+++ controllers/brandController.js
@@ -4,46 +4,63 @@ class BrandController {
   async getAllBrands(req, res) {
     try {
       const brands = await BrandService.getAll();
-      return res.json(brands);
+      res.json(brands);
     } catch (err) {
-      return res.status(500).json(err);
+      res.status(500).json(err);
     }
   }
   async getOneBrand(req, res) {
     try {
       const { id } = req.params;
       const brand = await BrandService.getOne(id);
-      return res.json(brand);
+      if (!brand) {
+        return res
+          .status(400)
+          .json({ message: 'Brand with this id doesnt exist' });
+      }
+      res.json(brand);
     } catch (err) {
-      return res.status.json(err);
+      res.status(500).json(err);
     }
   }
   async createBrand(req, res) {
     try {
       const { name } = req.body;
       const brand = await BrandService.create(name);
-      return res.json(brand);
+      res.json(brand);
     } catch (err) {
-      return res.status(500).json(err);
+      res.status(500).json(err.message);
     }
   }
   async deleteBrand(req, res) {
     try {
       const { id } = req.params;
+      const candidate = await BrandService.getOne(id);
+      if (!candidate) {
+        return res
+          .status(400)
+          .json({ message: 'Brand with this id doesnt exist' });
+      }
       const brand = await BrandService.delete(id);
-      return res.json(brand);
+      res.json(brand);
     } catch (err) {
-      return res.status(500).json(err);
+      res.status(500).json(err.message);
     }
   }
   async updateBrand(req, res) {
     try {
       const { id } = req.params;
       const { name } = req.body;
+      const candidate = await BrandService.getOne(id);
+      if (!candidate) {
+        return res
+          .status(400)
+          .json({ message: 'Brand with this id doesnt exist' });
+      }
       const brand = await BrandService.update(id, name);
-      return res.json(brand);
+      res.json(brand);
     } catch (err) {
-      return res.status(500).json(err);
+      res.status(500).json(err.message);
     }
   }
 }
